Create API endpoint to read table data
Move seed code into seperate seed.js file
Change index.js to only select/read tables
